pyproject.toml
[tool.poetry]
name = "Inventory Monitor API"
version = "0.0.1"
description = "A tool to monitor for when items become available "
authors = ["James Tombleson <luther38@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.70.0"
selenium = "^4.0.0"
APScheduler = "^3.8.0"
uvicorn = "^0.15.0"
python-dotenv = "^0.19.1"
SQLAlchemy = "^1.4.26"
FastAPI-SQLAlchemy = "^0.2.1"
beautifulsoup4 = "^4.10.0"
alembic = "^1.7.4"
psycopg2-binary = "^2.9.1"
pyroscope-io = "^0.5.1"
fastapi-healthcheck = "^0.2.1"
fastapi-healthcheck-sqlalchemy = "^0.1.0"
fastapi-healthcheck-uri = "^0.1.0"

[tool.poetry.dev-dependencies]
mypy = "^0.910"
flake8 = "^4.0.1"
pytest = "^6.2.5"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
